core/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
prisma>=0.11.0

# MQTT
paho-mqtt>=1.6.1

# WebSocket support is included in FastAPI/Uvicorn
